+++ edge_to_graph.c
@@ -52,9 +52,9 @@ int main(int argc, char *argv[]) {
 		exit(EXIT_FAILURE);
 	}
 
-	unsigned long PAGE_SIZE = sysconf(_SC_PAGESIZE);
+	int PAGE_SIZE = sysconf(_SC_PAGESIZE);
 
-	unsigned long header_size = sizeof(unsigned long)*(3 + 3*N); //N, M, D(3) [o,d]*N in 
+	unsigned long header_size = sizeof(int)*(3 + 3*N); //N, M, D(3) [o,d]*N in 
 	printf("The header is (%d x 2 + 5) X 4 = %d bytes\n", N, header_size);
 	printf("The pages are %d bytes\n", PAGE_SIZE);
 	
@@ -75,18 +75,18 @@ int main(int argc, char *argv[]) {
 
 	printf("The upper bound of D is %d\n", D_up);
 	
-	unsigned long node_data_size = sizeof(unsigned long)*N*D_up; //in bytes
+	unsigned long node_data_size = sizeof(int)*N*D_up; //in bytes
 
 	
 	//stretch the file in bytes
-	size_t length = num_hpages*PAGE_SIZE + sizeof(unsigned long)*D_up + node_data_size; // N, M, D; [n,o]*N + padding; burn row [N][D]
+	size_t length = num_hpages*PAGE_SIZE + sizeof(int)*D_up + node_data_size; // N, M, D; [n,o]*N + padding; burn row [N][D]
 	if (lseek(fd, length+1, SEEK_SET) == -1){
 		close(fd);
 		perror("Error on lseek call to stretch the file");
 		exit(EXIT_FAILURE);
 	} //position at end of file
 	
-	double dpages = ((sizeof(unsigned long)*D_up + node_data_size))/(double)PAGE_SIZE;
+	double dpages = ((sizeof(int)*D_up + node_data_size))/(double)PAGE_SIZE;
 	printf("There are %f double node pages\n", dpages);
 	
 	unsigned long num_dpages = ceil(dpages);
@@ -94,7 +94,7 @@ int main(int argc, char *argv[]) {
 	
 	unsigned long num_tpages = num_hpages + num_dpages;
 
-	printf("The length of this file is based on header %d + padding %d + node data %d + burned row %d = %d\n", header_size, padding, node_data_size, sizeof(unsigned long)*D_up, length);
+	printf("The length of this file is based on header %d + padding %d + node data %d + burned row %d = %d\n", header_size, padding, node_data_size, sizeof(int)*D_up, length);
 	
 	printf("The length of the file is %d, requiring %d pages\n", length, num_tpages);
 
@@ -121,20 +121,20 @@ int main(int argc, char *argv[]) {
 
 	rewind(fp);
 	
-	unsigned long buf[3]={(unsigned long)N, (unsigned long)0, (unsigned long)D_up};
+	int buf[3]={(int)N, (int)0, (int)D_up};
 	
-	fwrite(buf, sizeof(unsigned long), 3, fp);
+	fwrite(buf, sizeof(int), 3, fp);
 	
 	
 	rewind(fp);
 	
 	
-	unsigned long c;
-	c=fread(buf, sizeof(unsigned long), 1, fp);
+	int c;
+	c=fread(buf, sizeof(int), 1, fp);
 	
-	c=fread(buf, sizeof(unsigned long), 1, fp);
+	c=fread(buf, sizeof(int), 1, fp);
 		
-	c=fread(buf, sizeof(unsigned long), 1, fp);
+	c=fread(buf, sizeof(int), 1, fp);
 	
 	struct graph* my_graph = Graph(fd);
 	
@@ -146,7 +146,7 @@ int main(int argc, char *argv[]) {
 	}
 	
 	//initialize nodes with offset and degree 0
-	for(unsigned long i=0; i<my_graph->N ;i++){
+	for(int i=0; i<my_graph->N ;i++){
 		my_graph->map[3+2*i]=i;
 		my_graph->map[3+2*i+1]=0;
 		my_graph->map[3+2*my_graph->N+i]=i+1;
@@ -154,7 +154,7 @@ int main(int argc, char *argv[]) {
 
 	//print_graph(my_graph);
 	
-	unsigned long u,v;
+	int u,v;
 	while(fscanf(fe, "%d %d", &u, &v)==2){
 		//printf("Adding edge %d %d in e2g\n", u, v);
 		add_edge(my_graph, u, v);
